% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_knn_cv.R
\name{my_knn_cv}
\alias{my_knn_cv}
\title{K nearest neighbors approximation}
\usage{
my_knn_cv(train, cl, k_nn, k_cv)
}
\arguments{
\item{train}{An inputted data frame of numerics to act as training data.}

\item{cl}{An inputted vector of the true class value of our training data.}

\item{k_nn}{The number of neighbors to based our training example from.}

\item{k_cv}{The number of fold we wish to split our data into.}
}
\value{
A vector of predicted classes for our n-row observations, as well as
a single numeric averaged misclassification error from the number of
neighbors \code{k_nn}.
}
\description{
This function approximates a k-nearest neighbors algorithm for a data frame.
}
\examples{
# Gets rid of NA values from penguins data
penguins <- na.omit(NewPackageLecture9::my_penguins)

# Creates data frame of just numeric values
pen_num <- penguins \%>\% dplyr::select(bill_length_mm, bill_depth_mm, flipper_length_mm, body_mass_g)

# Creates vector of the species classifications
pen_cl <- penguins \%>\% dplyr::pull(species)

# Makes vector of strings of those classes
pen_cl <- as.character(pen_cl)

my_knn_cv(pen_num, pen_cl, 5, 5)
# my_knn_cv(NUMERIC_df, CLASS_vector, 3, 5)

}
\keyword{and}
\keyword{inference}
\keyword{prediction}
