% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_lm.R
\name{my_lm}
\alias{my_lm}
\title{Linear model maker}
\usage{
my_lm(formula, data)
}
\arguments{
\item{formula}{A formula class object that describes which two numerics  will
be used to calculate the linear model.}

\item{data}{An inputted data frame of numerics.}
}
\value{
Table of four variables, including the estimated linear regression
coefficient \code{Estimates}, the standard error \code{std.Error}, the
t-value \code{t value}, and the p-value \code{Pr(>|t|)}.
}
\description{
This function calculates a linear model.
}
\examples{
penguins <- na.omit(NewPackageLecture9::my_penguins)

my_lm(body_mass_g ~ bill_length_mm + bill_depth_mm, penguins) #using the
  # palmerspenguins data that is included in this package
# my_lm(NUMERIC_VAR_3 ~ NUMERIC_VAR_1 + NUMERIC_VAR_2, DATA_FRAME)

}
\keyword{inference}
