% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_t.test.R
\name{my_t.test}
\alias{my_t.test}
\title{T-test maker}
\usage{
my_t.test(x, alternative, mu)
}
\arguments{
\item{x}{Vector input of numerics.}

\item{alternative}{String input of either "less", "greater", or "two.sided"
which tells the function \code{my_t.test} which side of the area under the
curve to calculate.}

\item{mu}{Single numeric value which is the mean of our test population.}
}
\value{
List of a data frame representing the numeric test statistic \code{test_stat},
the numeric degrees of freedom \code{df}, a string indicating which side of
the model we were calculating \code{alternative}, and the numeric p value \code{p_val}.
}
\description{
This function preforms a t-test on a vector of data.
}
\examples{
my_t.test(1:15, "less", 0)
my_t.test(c(10,11,12), "two.sided", 11)

}
\keyword{inference}
